Separate LSPS2 schema's
Separate the lsps2-schema's from the message file
